+++ bin/orbis-node/src/sign/mod.rs
@@ -1,5 +1,6 @@
 pub mod coordinator;
 pub mod error;
+pub mod helpers;
 pub mod messages;
 pub mod protocol_handler;
 pub mod response_state;
